Fix a race condition in the expandable transition code
When fixing a layout bug in cf-expandables it turned out there was a
race condition causing the layout for the stored min-height to be
calculated prior to the final component rendering. The changes here add
an onLoad utility to store scripting that must run after the page is
ready and run it at that point.