Transpile source to lib/ dir for npm publishing
By pointing package.json's `main` property to the ES6 source code,
external projects that use this module have to ensure they transpile
the code before publishing it. It's better practice to pre-transpile
everything to a lib/ dir and set it as `main` so that other projects
can consume it hassle-free.

https://stackoverflow.com/a/33976278